modify code

+++ src/main/java/com/demo/springsecurityblogappmuhammedduzgun/service/PostService.java
@@ -1,14 +1,15 @@
 package com.demo.springsecurityblogappmuhammedduzgun.service;
 
 import com.demo.springsecurityblogappmuhammedduzgun.entity.Post;
+import com.demo.springsecurityblogappmuhammedduzgun.entity.User;
 import com.demo.springsecurityblogappmuhammedduzgun.exception.EntityNotFoundException;
 import com.demo.springsecurityblogappmuhammedduzgun.mapper.PostMapper;
 import com.demo.springsecurityblogappmuhammedduzgun.repository.PostRepository;
 import com.demo.springsecurityblogappmuhammedduzgun.request.post.CreatePostRequest;
 import com.demo.springsecurityblogappmuhammedduzgun.response.post.PostResponse;
+import com.demo.springsecurityblogappmuhammedduzgun.security.BlogUserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 import java.util.UUID;
 
@@ -25,8 +26,9 @@ public class PostService {
 
     //Create Post
     @Transactional
-    public PostResponse createPost(CreatePostRequest createPostRequest) {
+    public PostResponse createPost(CreatePostRequest createPostRequest, BlogUserDetails blogUserDetails) {
         Post postToCreate = postMapper.mapToPost(createPostRequest);
+        postToCreate.setUser(blogUserDetails.getUser());
         Post createdPost = postRepository.save(postToCreate);
         return postMapper.mapToPostResponse(createdPost);
     }
@@ -46,7 +48,15 @@ public class PostService {
 
     //Delete Post By Id
     @Transactional
-    public void deletePostById(UUID id) {
+    public void deletePostById(BlogUserDetails userDetails, UUID id) {
+        User user = userDetails.getUser();
+        //is post exists
+        Post postToDelete = postRepository.findById(id)
+                        .orElseThrow(()-> new EntityNotFoundException("Post with id " + id + " not found"));
+        //is post belongs to user
+        if (!postToDelete.getUser().getId().equals(user.getId())) {
+            throw new RuntimeException("this post does not belong to user");
+        }
         postRepository.deleteById(id);
     }
